FullSizeMainView done
