+++ WhiteNFluffyTestTask/CitiesList/CitiesListMainView.swift
